Reorganize backend types
