+++ backend/src/types/errorTypes.ts
@@ -1,0 +1,8 @@
+export class ApiError extends Error {
+  statusCode: number;
+
+  constructor(statusCode: number, message: string | undefined) {
+    super(message);
+    this.statusCode = statusCode;
+  }
+}
