+++ backend/src/types/ontologyTypes.ts
@@ -36,12 +36,3 @@ export type Annotation = {
   label: string;
   description: string;
 };
-
-export class ApiError extends Error {
-  statusCode: number;
-
-  constructor(statusCode: number, message: string | undefined) {
-    super(message);
-    this.statusCode = statusCode;
-  }
-}
